Solved Problem 3 after break

+++ Euler_Project/src/util/findPrime.c
@@ -1,8 +1,9 @@
 #include "findPrime.h"
 
 
-bool setPrimeList(int * prime_list, int array_length, int start, int end)
+bool setPrimeList(int * prime_list, int array_length, long start, long end)
 {
+    /// start and end are set as uint_64
     /// If list of primes is longer than array_length return false as an error.
 
     bool is_within_bounds = true;
@@ -11,8 +12,8 @@ bool setPrimeList(int * prime_list, int array_length, int start, int end)
 }
 
 
-int findNextPrimeNumber(int current_prime_number){
-    int next_prime_number = -1;
+long findNextPrimeNumber(long current_prime_number){
+    long next_prime_number = -1;
     bool not_prime = true;
     int increment = 1;
     // while (not_prime)
@@ -23,14 +24,14 @@ int findNextPrimeNumber(int current_prime_number){
     return next_prime_number;
 }
 
-bool brute_force_prime(int integer)
+bool brute_force_prime(long integer)
 {
     if (isPrime(integer))
         return true;
     double int_squared_root = sqrt(integer-1);
     /// brute force check
-    for (int j=7; j <= (int) int_squared_root+1; j++)
-        for (int k=1; k <= (int) int_squared_root+1; k++)
+    for (long j=7; j <= (long) int_squared_root+1; j++)
+        for (long k=1; k <= (long) int_squared_root+1; k++)
             {
                 if (integer % (j*k) == 0)
                     return false;
@@ -38,7 +39,7 @@ bool brute_force_prime(int integer)
     return true;
 }
 
-bool isPrime(int integer)
+bool isPrime(long integer)
 {
     // Trivial Case
     bool condition = false;
@@ -49,26 +50,26 @@ bool isPrime(int integer)
         || integer == 13
         )
         return true;
-    int test_int = integer % 10;
+    short test_int = integer % 10;
     if (abs((test_int % 2)-1))
         return condition;
     if ((test_int % 5) == 0)
         return condition;
     
     // Build Digit Stack
-    int log_10_power = 0;
+    long log_10_power = 0;
     for (int i=0; pow(10,log_10_power)<integer; i++)
         log_10_power = i;
     stack_handle digit_stack;
     initializeStack(&digit_stack);
-    int running_int = integer;
+    long running_int = integer;
 
     // Collect Digits into Digit Stack
     for (int i=1; i<log_10_power; i++)
     {
-        test_int = (running_int / (int) pow(10,i-1)) % (int) pow(10,i); 
+        test_int = (running_int / (long) pow(10,i-1)) % (long) pow(10,i); 
         digit_stack.push(&digit_stack, test_int);
-        running_int -= (test_int * (int) pow(10,i-1));
+        running_int -= (test_int * (long) pow(10,i-1));
     }
     // Process Digits Values
     running_int = 0;
@@ -80,7 +81,7 @@ bool isPrime(int integer)
         return condition;
     double int_squared_root = sqrt(integer-1);
     /// one more than perfect square theorem
-    if ((((int) int_squared_root) - int_squared_root) == 0)  
+    if ((((long) int_squared_root) - int_squared_root) == 0)  
         return true;
     condition = true;
