+++ Euler_Project/src/util/findPrime.h
@@ -2,16 +2,17 @@
 #define __PRIME_UTIL_H__
 
 #include "base.h"
+#include "stdint.h"
 
 #ifdef _WIN32
 #elif __linux__
 #endif
 
 
-bool setPrimeList(int * prime_list, int array_length, int start, int end);
-int findNextPrimeNumber(int current_prime_number);
-bool brute_force_prime(int integer);
-bool isPrime(int integer);
+bool setPrimeList(int * prime_list, int array_length, long start, long end);
+long findNextPrimeNumber(long current_prime_number);
+bool brute_force_prime(long integer);
+bool isPrime(long integer);
 
 
 #endif
