+++ Euler_Project/src/util/stack_set.h
@@ -9,7 +9,7 @@
 
 struct Stack
 {
-    int data;
+    long data;
     struct Stack * next_ptr;
 };
 
@@ -18,10 +18,10 @@ typedef struct Stack *stack_ptr;
 struct StackHandler
 {
     /// Class Variables
-    // int top_value;   // Top Stack Vector Data w/out popping stack
+    // long top_value;   // Top Stack Vector Data w/out popping stack
         //// TODO: top_value should not be callable; utilize a getter method
         //          in order to compensate for NULL Vector
-    int bottom;
+    long bottom;
     int stack_depth;
     short _shift_state;  // DEV PLACEHOLDEER 
 
@@ -30,12 +30,11 @@ struct StackHandler
     struct Stack * bottom_ptr;
     
     /// Class Method Vectors
-    int (*getTopValue)(struct StackHandler * this);
-    int (*pop)(struct StackHandler * this);
+    long (*getTopValue)(struct StackHandler * this);
+    long (*pop)(struct StackHandler * this);
     void (*shift)(struct StackHandler * this);
     void (*unshift)(struct StackHandler * this);  
-    void (*push)(struct StackHandler * this,
-                    int value);
+    void (*push)(struct StackHandler * this, long value);
 };
 typedef struct StackHandler stack_handle;
 
@@ -43,9 +42,9 @@ typedef struct StackHandler stack_handle;
 void buildFILOStack(struct Stack * this);
 void initializeStack(struct StackHandler * this);
 
-int __get_top_stack_value__(struct StackHandler * this);
-void __FILO_stack_push__(struct StackHandler * this, int value);
-int __FILO_stack_pop__(struct StackHandler * this);
+long __get_top_stack_value__(struct StackHandler * this);
+void __FILO_stack_push__(struct StackHandler * this, long value);
+long __FILO_stack_pop__(struct StackHandler * this);
 void __FILO_stack_shift__(struct StackHandler * this, bool is_to_shift_up);
 
 void __FILO_stack_shift_up__(struct StackHandler * this);
@@ -56,10 +55,10 @@ void __FILO_stack_shift_down__(struct StackHandler * this);
 struct InvStackHandler
 {
     /// Class Variables
-    // int top_value;   // Top Stack Vector Data w/out popping stack
+    // long top_value;   // Top Stack Vector Data w/out popping stack
         //// TODO: top_value should not be callable; utilize a getter method
         //          in order to compensate for NULL Vector
-    int top;
+    long top;
     int stack_depth;
     short _shift_state;  // DEV PLACEHOLDEER --> delete once testing is complete
 
@@ -69,21 +68,20 @@ struct InvStackHandler
     struct Stack * top_ptr;
     
     /// Class Method Vectors
-    int (*getBottomValue)(struct InvStackHandler * this);
-    int (*pop)(struct InvStackHandler * this);
+    long (*getBottomValue)(struct InvStackHandler * this);
+    long (*pop)(struct InvStackHandler * this);
     void (*shift)(struct InvStackHandler * this);
     void (*unshift)(struct InvStackHandler * this);  
-    void (*push)(struct InvStackHandler * this,
-                    int value);
+    void (*push)(struct InvStackHandler * this, long value);
 };
 
 
 void buildFIFOStack(struct Stack * this);
 void initializeInvStack(struct InvStackHandler * this);
 
-int __get_bottom_stack_value__(struct InvStackHandler * this);
-void __FIFO_stack_push__(struct InvStackHandler * this, int value);
-int __FIFO_stack_pop__(struct InvStackHandler * this);
+long __get_bottom_stack_value__(struct InvStackHandler * this);
+void __FIFO_stack_push__(struct InvStackHandler * this, long value);
+long __FIFO_stack_pop__(struct InvStackHandler * this);
 void __FIFO_stack_shift__(struct InvStackHandler * this, bool is_to_shift_down);
 
 void __FIFO_stack_shift_down__(struct InvStackHandler * this);
